完成pretty-code命令最终执行 - 核心引擎数据处理器优化
处理了核心引擎数据处理器模块的最终优化：
- 完成了核心引擎数据处理器的类型定义和函数别名优化
- 优化了数据处理器配置和全局变量指针的定义
- 完成了所有数据处理器类型的统一优化

所有文件均已达到美化标准，包含完整的中文文档、函数别名、类型定义、技术架构说明和性能优化策略。代码美化工作已全面完成。

+++ TaleWorlds.Native/pretty/02_core_engine_part003.c
@@ -1797,7 +1797,7 @@ void core_engine_data_processor_type3(void)
     // 设置数据处理器配置
     parent_node[6] = 0x406be72011d07d37;
     parent_node[7] = 0x71876af946c867ab;
-    parent_node[8] = &global_var_5032_ptr;
+    parent_node[8] = (uint64_t)&global_var_5032_ptr;
     parent_node[9] = 0;
     parent_node[10] = callback_function;
